refactor(Operators): rename string to toString
