+++ packages/cerebral/src/operators/index.js
@@ -5,7 +5,7 @@ export {default as equals} from './equals'
 
 export {default as state} from './state'
 export {default as input} from './input'
-export {default as string} from './string'
+export {default as toString} from './toString'
 
 export {default as concat} from './concat'
 export {default as merge} from './merge'
