+++ packages/cerebral/src/operators/toString.js
@@ -1,6 +1,6 @@
 import populatePath from './helpers/populatePath'
 
-export default function string (strings, ...values) {
+export default function toString (strings, ...values) {
   return (context) => {
     const target = 'string'
     const path = populatePath(context, strings, values)
