+++ packages/tutorial/parts/08/src/index.js
@@ -5,7 +5,7 @@ import App from './components/App'
 import {Container} from 'cerebral/react'
 import Devtools from 'cerebral/devtools'
 import HttpProvider from 'cerebral-provider-http'
-import {set, state, debounce, input, string} from 'cerebral/operators'
+import {set, state, debounce, input, toString} from 'cerebral/operators'
 
 const toastDebounce = debounce.shared()
 function showToast (message, ms, type = null) {
@@ -58,7 +58,7 @@ const controller = Controller({
         }
       ],
       setStarsSum,
-      ...showToast(string`The repos has a total star count of ${state`starsSum`}`, 4000, 'success')
+      ...showToast(toString`The repos has a total star count of ${state`starsSum`}`, 4000, 'success')
     ]
   },
   providers: [
